moved and relinked folder structure in backend
