+++ api/battleship/events/events.py
@@ -1,6 +1,6 @@
 from flask_socketio import SocketIO, emit, join_room, leave_room, send
 from flask import request, session
-from .extensions import socketio
+from ..utils.extensions import socketio
 
 ROOMS = {}
