+++ api/battleship/routes/auth.py
@@ -5,7 +5,7 @@ from flask import (
 )
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from . import db
+from ..database import db
 
 bp = Blueprint('auth', __name__, url_prefix='/auth')
