+++ api/battleship/utils/extensions.py
